Interactive and non interactive mode setup
